requirements.txt
flask
torch
torchvision
onnxruntime
openvino
numpy
Pillow
opencv-python
tensorrt
pyqt5
requests